+++ marklogic-langchain4j/build.gradle
@@ -1,18 +1,14 @@
+/**
+ * This project defines processing that depends on langchain4j and thus will require Java 17.
+ */
+
 java {
-  // Must use Java 17 as langchain4j depends on 17+.
   toolchain {
     languageVersion = JavaLanguageVersion.of(17)
   }
 }
 
-/**
- * This project defines processing that depends on langchain4j and thus will require Java 17.
- */
 dependencies {
   api project(":marklogic-spark-api")
-
-  compileOnly "org.apache.spark:spark-sql_2.12:" + sparkVersion
-
-  // For splitting text and creating embeddings.
   api "dev.langchain4j:langchain4j:1.0.0-alpha1"
 }
