+++ marklogic-spark-api/src/main/java/com/marklogic/spark/core/embedding/ChunkSelector.java
@@ -3,7 +3,6 @@
  */
 package com.marklogic.spark.core.embedding;
 
-import com.marklogic.client.document.DocumentWriteOperation;
 import com.marklogic.client.io.marker.AbstractWriteHandle;
 
 /**
@@ -12,12 +11,10 @@ import com.marklogic.client.io.marker.AbstractWriteHandle;
 public interface ChunkSelector {
 
     /**
-     * @param sourceDocument
+     * @param uri
+     * @param content
      * @return Selecting chunks may involve deserializing a string or byte array into an e.g. JsonNode, in which case
      * the document to be written will not be the sourceDocument that is passed in.
      */
-    DocumentAndChunks selectChunks(DocumentWriteOperation sourceDocument);
-
-    // The new way, above method will be removed soon.
     DocumentAndChunks selectChunks(String uri, AbstractWriteHandle content);
 }
